problem 09: re-implement parameter handling
Ran into bugs with Problem 9 Part 1 with handling of different parameter
modes; the bugs had to do with mixing logic regarding the modes in the
individual op methods and not differentiating between dereferencing a
parameter in order to read a value from memory versus finding a location
to write to.

Added new methods `read_param(param_mode)` and `write_param(param_mode,
value)` which encapsulate all of the logic around reading a value from
memory based on the given parameter, and writing a value to memory based
on a given parameter - including checking the parameter mode, and
writing to / reading from memory.