finish problem 12 part two
initial approach was to keep track of the history of the moon's
positions and velocities after each simulation round by converting the
list of moons to a string and putting the string in a set. But this
takes forever to run.

After accidentally reading some hints, realize that each X/Y/Z
coordinate is independent, so we can find the cycle length for each
coordinate, then take the least common multiple of the cycle lengths to
find the cycle length of the whole system.