problem 15 part 1: done
fixed a big bug where after the robot found the oxygen station,
computing the shortest path from the starting position to the oxygen
station included moving into UNKNOWN tiles on the map. Allowing this
makes sense when the destination is an UNKNOWN tile, but not for when we
want to find the path for a known destination. Added a keyword parameter
to alter the behavior in this mode.