vastly simplify problem 15, solve part two
cut runtime down from 30-60 seconds to 1-2 by doing a DFS search to
explore the whole map, then counting the number of moves from the
starting position to the goal, and the goal to the furthest spot on
board, is a simple BFS search. This is sped up by removing all of the
distance calculations (of the entire board) on each move.